Change the way scripts are loaded
