+++ pages/_app.tsx
@@ -2,7 +2,6 @@ import { useRouter } from 'next/router'
 import { useEffect } from 'react'
 import type { AppProps } from 'next/app'
 import * as gtag from '../utils/gtag'
-import Script from 'next/script'
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
 
@@ -13,7 +12,7 @@ import Layout from '../components/layout/Layout'
 import PageTransition, { useAsPathWithoutHash } from '@madeinhaus/nextjs-page-transition'
 import '@madeinhaus/nextjs-page-transition/dist/index.css'
 
-function MyApp({ Component, pageProps }: AppProps) {
+function App({ Component, pageProps }: AppProps) {
 
   const key = useAsPathWithoutHash()
   const router = useRouter()
@@ -31,24 +30,6 @@ function MyApp({ Component, pageProps }: AppProps) {
   }, [router.events])
 
   return <>
-    <Script
-      strategy="lazyOnload"
-      src='https://www.googletagmanager.com/gtag/js?id=G-NHDRDGHWQ9'
-    />
-    <Script
-      id='google-analytics'
-      strategy="lazyOnload"
-      dangerouslySetInnerHTML={{
-        __html: `
-          window.dataLayer = window.dataLayer || [];
-          function gtag(){dataLayer.push(arguments);}
-          gtag('js', new Date());
-          gtag('config', 'G-NHDRDGHWQ9', {
-            page_path: window.location.pathname,
-          });
-        `
-      }}
-    />
     <Layout>
       <PageTransition
         as='div'
@@ -61,4 +42,4 @@ function MyApp({ Component, pageProps }: AppProps) {
   </>
 }
 
-export default MyApp
+export default App
